admin role

+++ app/Http/Controllers/CategoryController.php
@@ -4,11 +4,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Category;
+use Illuminate\Support\Facades\Session;
 
 class CategoryController extends Controller
 {
     public function addCategory(Request $request)
     {
+        if(Session::get('adminDetails')['category_edit_access'] == 0)
+        {
+            return redirect('/admin/dashboard')->with('flash_message_error','You have no access to this model');
+        }
         if($request->isMethod('post')){
             $data = $request->all();
 
@@ -37,12 +42,22 @@ class CategoryController extends Controller
 
     public function view_category()
     {
+        if(Session::get('adminDetails')['category_view_access'] == 0)
+        {
+            return redirect('/admin/dashboard')->with('flash_message_error','You have no access to this model');
+        }
+        
         $categories = Category::orderBy('id','DESC')->get();
         return view('admin.categories.view_category')->with(compact('categories'));
     }
 
     public function edit_category(Request $request, $id)
     {
+        if(Session::get('adminDetails')['category_edit_access'] == 0)
+        {
+            return redirect('/admin/dashboard')->with('flash_message_error','You have no access to this model');
+        }
+        
         if($request->isMethod('post')){
             $data = $request->all();
 
@@ -71,6 +86,10 @@ class CategoryController extends Controller
 
     public function delete_category($id)
     {
+        if(Session::get('adminDetails')['category_full_access'] == 0)
+        {
+            return redirect('/admin/dashboard')->with('flash_message_error','You have no access to this model');
+        }
         $category = Category::where('id',$id)->delete();
         return redirect()->back()->with('flash_message_success', 'Category Deleted Successfully');
     }
