+++ app/Http/Controllers/UsersController.php
@@ -62,7 +62,9 @@ class UsersController extends Controller
                 $user->name = $data['name'];
                 $user->email = $data['email'];
                 $user->password = bcrypt($data['password']);
-
+                date_default_timezone_set('Asia/Dhaka');
+                $user->created_at = date('Y-m-d H:i:s');
+                $user->updated_at = date('Y-m-d H:i:s');
                 $user->save();
     
                 //send welcome email
@@ -235,12 +237,20 @@ class UsersController extends Controller
     
     public function viewUsers()
     {
+        if(Session::get('adminDetails')['user_access'] == 0)
+        {
+            return redirect('/admin/dashboard')->with('flash_message_error','You have no access to this model');
+        }
         $users = User::all();
         return view('admin.users.view_users',compact('users'));
     }
     
     public function delete_user($id)
     {
+        if(Session::get('adminDetails')['user_access'] == 0)
+        {
+            return redirect('/admin/dashboard')->with('flash_message_error','You have no access to this model');
+        }
         $user = User::where('id',$id)->delete();
         return redirect()->back()->with('flash_message_success', 'User Deleted Successfully');
     }
